requirements.txt
cryptography
libnacl
aiohttp==3.8.6; python_version=='3.7'
aiohttp>=3.9.1; python_version>'3.7'
aiohttp_apispec>=3.0.0b1
pyOpenSSL
pyasn1
asynctest; python_version=='3.7'
marshmallow
typing-extensions
packaging
apispec>=6.0.0
pyipv8
streamlit
